Fixed a bug in get_threshold.
More info:-
The bug is in the division operation in the return
statement. The division is returning 0 since the two
operands are integers.

+++ hw2/hw2/PercolationStats.java
@@ -26,7 +26,7 @@ public class PercolationStats {
             p.open(row, col);
         }
 
-        return p.numberOfOpenSites() / (N * N);
+        return (double)(p.numberOfOpenSites()) / (N * N);
 
     }
